Added AlignBase code and refactor classes that could use it
