+++ robot.py
@@ -38,7 +38,6 @@ class Robot(commands2.TimedCommandRobot):
         """
         self.drivetrain.setDefaultCommand(DriveField(self.drivetrain, self.xbox_controller))
 
-
         """
         Setups
         """
@@ -61,7 +60,7 @@ class Robot(commands2.TimedCommandRobot):
         """
         Send commands to dashboard to
         """
-        pass
+        putCommandOnDashboard("Drivetrain", Drive(self.drivetrain, self.xbox_controller))
 
     def autonomousInit(self):
         self.auto_command: commands2.Command = self.auto_chooser.getSelected()
